fix(typehint): use redundant alias imports to fix mypy errors
PEP 484 requires redundant alias imports to identify modules as being re-exported. Adding these fixes the mypy error that it cannot find parent methods in child instances of `MainWindow`.

Fixes Issue #163
