+++ tests/test_qt.py
@@ -1,19 +1,30 @@
+import importlib
+import sys
+
 import pytest
 
+import pyvistaqt
 
-def test_no_qt_binding(no_qt):
-    from pyvistaqt import (
-        BackgroundPlotter,
-        MainWindow,
-        MultiPlotter,
-        QtInteractor,
-    )
 
-    with pytest.raises(RuntimeError, match="No Qt binding"):
-        BackgroundPlotter()
-    with pytest.raises(RuntimeError, match="No Qt binding"):
-        MainWindow()
-    with pytest.raises(RuntimeError, match="No Qt binding"):
-        MultiPlotter()
-    with pytest.raises(RuntimeError, match="No Qt binding"):
-        QtInteractor()
+def _check_qt_installed():
+    try:
+        from qtpy import QtCore  # noqa
+    except Exception:
+        return False
+    else:
+        return True
+
+
+def test_no_qt_binding(monkeypatch):
+    need_reload = False
+    if _check_qt_installed():
+        need_reload = True
+        monkeypatch.setenv('QT_API', 'bad_name')
+        sys.modules.pop('qtpy')
+        assert 'qtpy' not in sys.modules
+        with pytest.raises(RuntimeError, match="No Qt binding"):
+            importlib.reload(pyvistaqt)
+    monkeypatch.undo()
+    if need_reload:
+        importlib.reload(pyvistaqt)
+        assert 'qtpy' in sys.modules
